LayoutTests/imported/w3c/web-platform-tests/css/css-color-adjust/meta/meta-color-scheme-single-value-in-head-expected.txt


FAIL Meta color-scheme in head applies. assert_equals: expected "dark" but got "auto"
